+++ src/core/tradingEngineService/src/test/java/integrationTests/ContextRehydrationTests.java
@@ -9,6 +9,8 @@ import shared.service.TradingEngineServiceConsts;
 import shared.service.TradingEngineServiceRequest;
 import tradingEngineService.tradingEngine.TradingEngineContext;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ContextConfiguration(initializers = ContextRehydrationTests.ContextSeeder.class)
@@ -18,13 +20,13 @@ public class ContextRehydrationTests extends IntegrationTest {
     public void Test1() throws Exception {
         var request = TradingEngineServiceRequest.getBalance(TestHelpers.testAccount1.accountId(), TestHelpers.USD.code());
         var response = testClient.send(TradingEngineServiceConsts.WriteRequestTopic, "123", request);
-        assertEquals(100, response.getBalanceResult().balance());
+        assertEquals(100, response.getBalanceResult().balance().longValue());
     }
 
     public static class ContextSeeder extends AwsTradingEngineContextSeeder {
         @Override
         protected void seedContext(TradingEngineContext tradingEngineContext) {
-            tradingEngineContext.adjustBalance(TestHelpers.testAccount1.accountId(), TestHelpers.USD, 100);
+            tradingEngineContext.adjustBalance(TestHelpers.testAccount1.accountId(), TestHelpers.USD, BigDecimal.valueOf(100));
         }
     }
 }
