+++ src/core/tradingEngineService/src/test/java/integrationTests/ServiceTests.java
@@ -8,6 +8,8 @@ import shared.service.TradingEngineServiceConsts;
 import shared.service.TradingEngineServiceErrors;
 import shared.service.TradingEngineServiceRequest;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 public class ServiceTests extends IntegrationTest {
@@ -16,7 +18,7 @@ public class ServiceTests extends IntegrationTest {
 
     @Test
     public void orderShouldReject() throws Exception {
-        var request = TradingEngineServiceRequest.limitOrder(10, 100, 1, "SPY", OrderBookSide.Buy);
+        var request = TradingEngineServiceRequest.limitOrder(BigDecimal.valueOf(10), BigDecimal.valueOf(100), 1, "SPY", OrderBookSide.Buy);
         var response = testClient.send(TradingEngineServiceConsts.WriteRequestTopic, "123", request);
 
         assertNotNull(response.limitOrderResult());
@@ -25,30 +27,30 @@ public class ServiceTests extends IntegrationTest {
 
     @Test
     public void adjustBalance() throws Exception {
-        send(TradingEngineServiceRequest.adjustBalance(10, 1, "USD"));
+        send(TradingEngineServiceRequest.adjustBalance(BigDecimal.valueOf(10), 1, "USD"));
         var response = send(TradingEngineServiceRequest.getBalance(1, "USD"));
-        assertEquals(10, response.getBalanceResult().balance());
+        assertEquals(10, response.getBalanceResult().balance().longValue());
     }
 
     @Test
     public void placeOrder() throws Exception {
-        send(TradingEngineServiceRequest.adjustBalance(1000, 1, "USD"));
-        var response = send(TradingEngineServiceRequest.limitOrder(1, 100, 1, "SPY", OrderBookSide.Buy));
+        send(TradingEngineServiceRequest.adjustBalance(BigDecimal.valueOf(1000), 1, "USD"));
+        var response = send(TradingEngineServiceRequest.limitOrder(BigDecimal.valueOf(1), BigDecimal.valueOf(100), 1, "SPY", OrderBookSide.Buy));
         assertEquals(LimitOrderResultStatus.Ok, response.limitOrderResult().type());
     }
 
     @Test
     public void cancelOrder() throws Exception {
-        send(TradingEngineServiceRequest.adjustBalance(1000, 1, "USD"));
-        var response = send(TradingEngineServiceRequest.limitOrder(1, 100, 1, "SPY", OrderBookSide.Buy));
+        send(TradingEngineServiceRequest.adjustBalance(BigDecimal.valueOf(1000), 1, "USD"));
+        var response = send(TradingEngineServiceRequest.limitOrder(BigDecimal.valueOf(1), BigDecimal.valueOf(100), 1, "SPY", OrderBookSide.Buy));
         assertEquals(LimitOrderResultStatus.Ok, response.limitOrderResult().type());
         var orderId = response.limitOrderResult().result().restingOrder().id();
         response = send(TradingEngineServiceRequest.getBalance(1, "USD"));
-        assertEquals(900, response.getBalanceResult().balance());
+        assertEquals(900, response.getBalanceResult().balance().longValue());
         response = send(TradingEngineServiceRequest.cancel(1, "SPY", orderId));
         assertTrue(response.cancelOrderResult().success());
         response = send(TradingEngineServiceRequest.getBalance(1, "USD"));
-        assertEquals(1000, response.getBalanceResult().balance());
+        assertEquals(1000, response.getBalanceResult().balance().longValue());
     }
 
     @Test
@@ -59,18 +61,18 @@ public class ServiceTests extends IntegrationTest {
 
     @Test
     public void executeTrade() throws Exception {
-        send(TradingEngineServiceRequest.adjustBalance(1000, 1, "USD"));
-        send(TradingEngineServiceRequest.adjustBalance(1, 2, "SPY"));
-        var response = send(TradingEngineServiceRequest.limitOrder(1, 100, 1, "SPY", OrderBookSide.Buy));
+        send(TradingEngineServiceRequest.adjustBalance(BigDecimal.valueOf(1000), 1, "USD"));
+        send(TradingEngineServiceRequest.adjustBalance(BigDecimal.valueOf(1), 2, "SPY"));
+        var response = send(TradingEngineServiceRequest.limitOrder(BigDecimal.valueOf(1), BigDecimal.valueOf(100), 1, "SPY", OrderBookSide.Buy));
         assertEquals(LimitOrderResultStatus.Ok, response.limitOrderResult().type());
-        response = send(TradingEngineServiceRequest.limitOrder(1, 99, 2, "SPY", OrderBookSide.Sell));
+        response = send(TradingEngineServiceRequest.limitOrder(BigDecimal.valueOf(1), BigDecimal.valueOf(99), 2, "SPY", OrderBookSide.Sell));
         assertEquals(LimitOrderResultStatus.Ok, response.limitOrderResult().type());
         assertEquals(1, response.limitOrderResult().result().trades().size());
     }
 
     @Test
     public void adjustBalanceUnknownAsset() throws Exception {
-        var resp = send(TradingEngineServiceRequest.adjustBalance(10, 1, "XYZ"));
+        var resp = send(TradingEngineServiceRequest.adjustBalance(BigDecimal.valueOf(10), 1, "XYZ"));
         assertEquals(TradingEngineServiceErrors.UnknownAsset, resp.tradingEngineErrorResult().message());
     }
